Move files to columns and tables folders
