+++ packages/berlin/src/components/columns/results-columns/ResultsColumns.styled.tsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
-import { Grid } from '../containers/Grid.styled';
+import { Grid } from '../../containers/Grid.styled';
 
 export const Card = styled(Grid)`
   border-bottom: 2px solid var(--color-black);
   padding: 1.5rem;
-  grid-template-columns: auto 240px;
+  grid-template-columns: auto repeat(3, 48px) 80px;
 `;
