+++ packages/berlin/src/components/columns/stats-columns/StatsColumns.tsx
@@ -1,4 +1,4 @@
-import { Bold } from '../typography/Bold.styled';
+import { Bold } from '../../typography/Bold.styled';
 import { Card } from './StatsColumns.styled';
 
 function StatsColumns() {
