+++ packages/berlin/src/components/tables/comment-card/CommentCard.tsx
@@ -7,16 +7,16 @@ import { useParams } from 'react-router-dom';
 import { GetCommentsResponse, deleteComment, deleteLike, fetchCommentLikes, postLike } from 'api';
 
 // Hooks
-import useUser from '../../hooks/useUser';
+import useUser from '../../../hooks/useUser';
 
 // Store
-import { useAppStore } from '../../store';
+import { useAppStore } from '../../../store';
 
 // Components
-import { FlexRow } from '../containers/FlexRow.styled';
-import Button from '../button';
-import Dialog from '../dialog';
-import IconButton from '../icon-button';
+import { FlexRow } from '../../containers/FlexRow.styled';
+import Button from '../../button';
+import Dialog from '../../dialog';
+import IconButton from '../../icon-button';
 
 // Styled Components
 import { Author, Card, Comment, FormattedDate } from './CommentCard.styled';
