+++ packages/berlin/src/components/tables/result-card/index.ts
